# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CreatePromoCode:
    properties:
      discount:
        type: number
      discount_type:
        type: string
      order_limit_price:
        type: number
      promoCode_name:
        type: string
    type: object
...
  models.PromoCodePrimaryKey:
    properties:
      promoCode_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /promoCode:
    get:
      consumes:
      - application/json
      description: Get List PromoCode
      operationId: get_list_promoCode
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List PromoCode
      tags:
      - PromoCode
    post:
      consumes:
      - application/json
      description: Create PromoCode
      operationId: create_promoCode
      parameters:
      - description: CreatePromoCodeRequest
        in: body
        name: promoCode
        required: true
        schema:
          $ref: '#/definitions/models.CreatePromoCode'
      produces:
      - application/json
      responses:
        "201":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create PromoCode
      tags:
      - PromoCode
  /promoCode/{id}:
    delete:
      consumes:
      - application/json
      description: Delete PromoCode
      operationId: delete_promoCode
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: DeletePromoCodeRequest
        in: body
        name: promoCode
        required: true
        schema:
          $ref: '#/definitions/models.PromoCodePrimaryKey'
      produces:
      - application/json
      responses:
        "204":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete PromoCode
      tags:
      - PromoCode
    get:
      consumes:
      - application/json
      description: Get By ID PromoCode
      operationId: get_by_id_promoCode
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: promo_code
        in: query
        name: promo_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By ID PromoCode
      tags:
      - PromoCode
...
swagger: "2.0"